WebSpider/requirements.txt
async-timeout==4.0.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
Deprecated==1.2.13
html5lib==1.1
idna==3.3
importlib-metadata==4.8.3
packaging==21.3
pyparsing==3.0.9
python-dateutil==2.8.2
redis==4.3.2
requests==2.27.1
rq==1.10.1
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.1.1
urllib3==1.26.9
webencodings==0.5.1
wrapt==1.14.1
zipp==3.6.0
